added exitRawMode, created a variable to store original attributes for terminal, and added library for a function atexit()
There are tons of comments that are probably going to be more confusing than helpful to someone else reading this.
It should be apparent that my comments are really for myself, to understand on a deeper level what I am changing and why.
It's getting a little messy because of these comments, so if you are trying to understand, pay close attention to the number in pipes at the beginning of each comment. Again, this tells you what step I am on.
All of the added comments and changed code in this step will be connected with |3|.