We removed the software control flow flag IXON
The buttons Ctrl + S would suspend software and Ctrl + Q would resume it. This is unnecessary and still get sin our way.
Read comments |7| to learn more about this, but it has to do when we needed our devices to let the software 'catch up'. At least that's my crude understanding with an even more crude explanation.