Updated comments about CTRL keys working in Cygwin, and using Cygwin for testing form now on
It could be that I am doing something wrong, however, building this in VSC, using WSL, and having to deal with all of VSC's wonderful shortcuts and helpful features- They are problematic for what I am building.
A text editor using Linux, or Linux subsystems.
It would appear that things are actually working how I was trying to build it, and as the tutorial built it to use in Cygwin (I don't see this explicitly stated, but he mentions several cases where we make changes specficially for Cygwin), this is where I will do all testing.
The understanding is happening. This is what matters to me.

This commit is just for comments explaining this.